+++ FatGuysSpeak.Client/Pages/PorkChopPage.xaml.cs
@@ -1,0 +1,6 @@
+namespace FatGuysSpeak.Client.Pages;
+
+public partial class PorkChopPage : ContentPage
+{
+    public PorkChopPage() => InitializeComponent();
+}